Data update: complete climate data until May 2025
Script change: The criterion for setting the meta information of the database has been changed to set the time range of available data: For the border of the most recent available data the average completeness of a month was set from ">0.5" to ">0.9". The reason to make the change was that initially the new range was set to "until 06/2025" because the average completenes of this month was 0.73 (data processing on 24th June). The change has an effect on the month/year range of datasets displayed in the header and on the presets of the last 12 months to be displayed after the tool has been loaded.